Widen EntityNode owner mask to long and reject overflowing schemas
isOwner/isLinked were int while the masks they intersect are long, so indexed
and relation types past StructIndex 31 silently got no owner bit and their rows
outlived deleted entities. Ceiling is now 63, enforced when the schema is built.

+++ src/ECS/Index/Indexes/AbstractComponentIndex.cs
@@ -24,7 +24,7 @@ public abstract class AbstractComponentIndex
     internal  readonly  EntityStore     store;
     internal  readonly  ComponentType   componentType;
     internal  readonly  int             structIndex;
-    internal  readonly  int             indexBit;
+    internal  readonly  long            indexBit;
     internal            bool            modified;
     #endregion
     
@@ -45,7 +45,7 @@ public abstract class AbstractComponentIndex
         this.componentType  = componentType;
         structIndex         = componentType.StructIndex;
         var types           = new ComponentTypes(componentType);
-        indexBit            = (int)types.bitSet.l0;
+        indexBit            = types.bitSet.l0;
     }
 }
